Refactor Custom views into their own package 2
