+++ app/src/main/java/com/jeeps/charades/GameScore.java
@@ -6,10 +6,9 @@ import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
 import android.widget.Button;
 
-import com.jeeps.charades.model.CustomTextView;
+import com.jeeps.charades.views.CustomTextView;
 
 import java.util.ArrayList;
