+++ app/src/main/res/layout/content_game_score.xml
@@ -14,7 +14,7 @@
     tools:showIn="@layout/activity_game_score"
     android:background="@drawable/vintage_comic_simple">
 
-    <com.jeeps.charades.model.CustomTextView
+    <com.jeeps.charades.views.CustomTextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/message"
@@ -41,7 +41,7 @@
             android:orientation="vertical"
             android:weightSum="100">
 
-            <com.jeeps.charades.model.CustomTextView
+            <com.jeeps.charades.views.CustomTextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:text="Correct"
@@ -52,7 +52,7 @@
                 android:gravity="center_horizontal"
                 android:textColor="#000"/>
 
-            <com.jeeps.charades.model.CustomTextView
+            <com.jeeps.charades.views.CustomTextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:id="@+id/correct_score"
@@ -70,7 +70,7 @@
             android:orientation="vertical"
             android:weightSum="100">
 
-            <com.jeeps.charades.model.CustomTextView
+            <com.jeeps.charades.views.CustomTextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:text="Incorrect"
@@ -81,7 +81,7 @@
                 android:gravity="center_horizontal"
                 android:textColor="#000"/>
 
-            <com.jeeps.charades.model.CustomTextView
+            <com.jeeps.charades.views.CustomTextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:id="@+id/incorrect_score"
