+++ app/src/main/res/layout/content_play_game.xml
@@ -34,7 +34,7 @@
             android:adjustViewBounds="true"
             android:padding="20dp"/>
 
-        <com.jeeps.charades.model.CustomTextView
+        <com.jeeps.charades.views.CustomTextView
             android:layout_width="wrap_content"
             android:layout_height="10dp"/>
 
@@ -56,7 +56,7 @@
         android:layout_height="match_parent"
         android:layout_toLeftOf="@id/buttons_layout">
 
-        <com.jeeps.charades.model.CustomTextView
+        <com.jeeps.charades.views.CustomTextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:id="@+id/duration_counter"
@@ -65,7 +65,7 @@
             android:textColor="#fff"/>
 
 
-        <com.jeeps.charades.model.CustomTextView
+        <com.jeeps.charades.views.CustomTextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="center"
